check github postman scripts

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -12,8 +12,8 @@ import java.time.LocalDate;
 import java.util.HashSet;
 import java.util.Set;
 
-@NoArgsConstructor // Generates a no-argument constructor
-@AllArgsConstructor // Generates a constructor with all fields
+@NoArgsConstructor
+@AllArgsConstructor
 @Data
 @Builder
 public class User {
@@ -31,5 +31,5 @@ public class User {
     @PastOrPresent(message = "Дата рождения не может быть в будущем")
     private LocalDate birthday;
     @Builder.Default
-    private Set<Long> friends = new HashSet<>();
+    private Set<Integer> friends = new HashSet<>();
 }
